fix: post now delete file too

+++ back/src/controllers/postController.ts
@@ -1,3 +1,5 @@
+import fs from "fs";
+import path from "path";
 import { PrismaClient } from "@prisma/client";
 const prisma = new PrismaClient();
 
@@ -9,6 +11,7 @@ class postController {
       if (!req.user) return res.status(404).send("You must be logged in");
       if (!req.file) return res.status(404).send("You must provide a file");
       const { filename, originalname, size, mimetype } = req.file;
+      console.log(filename);
       const { title, soundtrackUrl } = req.body;
 
       const post = await prisma.post.create({
@@ -71,20 +74,53 @@ class postController {
   async deletePost(req: Request, res: Response) {
     try {
       const { id } = req.params;
-      const post = await prisma.post.delete({
+      console.log(id);
+      const post = await prisma.post.findUnique({
         where: {
           id: id,
         },
+        include: { file: true },
       });
-      return res.status(200).json(post);
+
+      if (post && post.file) {
+        console.log("Estou tentando deletar o arquivo:");
+        const filename = post.file.filename;
+        const filePath = path.join(__dirname, "..", "..", "uploads", filename);
+        console.log(filePath);
+
+        await prisma.file.delete({
+          where: {
+            id: post.file.id,
+          },
+        });
+        console.log("Deletei o arquivo no BD");
+
+        await prisma.post.delete({
+          where: {
+            id: id,
+          },
+          include: { file: true },
+        });
+        console.log("Deletei o Post");
+
+        fs.unlink(filePath, (err) => {
+          if (err) {
+            console.log("Error: " + err);
+            return res.status(500).send("Something went wrong");
+          } else {
+            console.log("Deletei o arquivo no servidor");
+            return res.status(200).json(post);
+          }
+        });
+      }
     } catch (err) {}
   }
-  /* async getFiles(req: Request, res: Response) {
+  async getFiles(req: Request, res: Response) {
     try {
       const file = await prisma.file.findMany();
       return res.status(200).json(file);
     } catch (err) {}
-  } */
+  }
 }
 
 export default new postController();
